Refactor: Extract Game feature into separate modules
- Created `feature/game/api`, `feature/game/data`, and `feature/game/presentation` modules.
- Moved `Game` interface and `GameComponent` to `feature/game/presentation`.
- Added build scripts (`build.gradle.kts`) for the new modules, including dependencies and namespaces.
- Created `GamePresentationModule.kt` in `feature/game/presentation` for Koin dependency injection setup for the `GameComponent.Factory`.
