+++ feature/game/data/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+    alias(libs.plugins.holdmaster.kmplib)
+    alias(libs.plugins.holdmaster.koin)
+}
+
+kotlin {
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.feature.game.api)
+            implementation(projects.core.base)
+            implementation(projects.core.utils)
+        }
+    }
+}
+
+android {
+    namespace = "ru.bortsov.holdmaster.feature.game.data"
+}
