+++ feature/game/presentation/src/commonMain/kotlin/ru/bortsov/holdmaster/feature/game/presentation/Game.kt
@@ -1,4 +1,4 @@
-package ru.bortsov.holdmaster.composeapp.decompose.feature.tabs.game
+package ru.bortsov.holdmaster.feature.game.presentation
 
 import com.arkivanov.decompose.ComponentContext
